requirements.txt
openai>=1.50.0
facebook-business>=19.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
